More fire stuff
Added a sound effect for fire and made shitty heat stacks so that the fire isn't just floating there. The sound effect is from a free package from the asset store, and I've gone ahead and included it in our folders, in case we still want more sound effects. Though, if they cause any issues in performance, it's best we get rid of 'em.